+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [1.2.1](https://github.com/p-acid/blog-template/compare/v1.2.0...v1.2.1) (2022-12-02)
+
+
+### Bug Fixes
+
+* **seo:** Move NextSeo to outside of Head ([a347092](https://github.com/p-acid/blog-template/commit/a3470925ce0a2b285cc42612d63422e9ae6ca9b3))
+
+## [1.2.0](https://github.com/p-acid/blog-template/compare/v1.1.1...v1.2.0) (2022-12-02)
+
+
+### Features
+
+* **SEO:** Add next-seo setting ([#8](https://github.com/p-acid/blog-template/issues/8)) ([f572ec7](https://github.com/p-acid/blog-template/commit/f572ec7073f4da163d1a23dac65e8b0d6e6b3d09))
+
+
+### Styling
+
+* **dark-theme:** Add PostItem dark mode style ([f572ec7](https://github.com/p-acid/blog-template/commit/f572ec7073f4da163d1a23dac65e8b0d6e6b3d09))
+
+
+### Configuration Settings
+
+* **default-seo:** Add default-seo by NextSeo ([f572ec7](https://github.com/p-acid/blog-template/commit/f572ec7073f4da163d1a23dac65e8b0d6e6b3d09))
+* **detail-seo:** Add SEO props in post detail page ([f572ec7](https://github.com/p-acid/blog-template/commit/f572ec7073f4da163d1a23dac65e8b0d6e6b3d09))
+* **package:** Add next-seo package ([f572ec7](https://github.com/p-acid/blog-template/commit/f572ec7073f4da163d1a23dac65e8b0d6e6b3d09))
+
 ## [1.1.1](https://github.com/p-acid/blog-template/compare/v1.1.0...v1.1.1) (2022-12-02)